Copia

+++ Proyecto_Licorera_Corchos.web/Data/Entities/Productos.cs
@@ -12,7 +12,7 @@ namespace Proyecto_Licorera_Corchos.web.Data.Entities
         public string Categoria { get; set; }
 
         [Required(ErrorMessage = "El campo '{0}' es requerido.")] // Indica que es un campo requerido
-        public decimal Precio { get; set; } // Se utiliza decimal para el tipo MONEY
+        public float Precio { get; set; } // Se utiliza decimal para el tipo MONEY
 
         [MaxLength(100, ErrorMessage = "El campo '{0}' debe tener máximo '{1}' carácteres")] // Valida que el campo no supere 100 caracteres
         public string Descripcion { get; set; }
